+++ src/main/java/net/geforcemods/securitycraft/blockentities/SecretHangingSignBlockEntity.java
@@ -113,7 +113,7 @@ public class SecretHangingSignBlockEntity extends HangingSignBlockEntity impleme
 	}
 
 	@Override
-	public void onOptionChanged(Option<?> option) {
+	public <T> void onOptionChanged(Option<T> option) {
 		level.sendBlockUpdated(worldPosition, getBlockState(), getBlockState(), 2);
 		ICustomizable.super.onOptionChanged(option);
 	}