+++ Changelog.md
@@ -19,6 +19,7 @@
 - API: The BlockState parameters in ICodebreakable's methods have been removed
 - API: New Option "EntityDataWrappedOption" that connects an EntityDataAccessor with an Option, and corresponding converter method "wrapForEntityData"
 - API: New method Option#getValueText for getting a textual representation of the option's value
+- API: ICustomizable#onOptionChanged now has a proper generic type
 - Fix: Trying to place a Panic Button on top of powdered snow crashes the game
 - Fix: Occasional crash when opening the inventory in creative mode in certain situations
 - Fix: Reinforced fence gates don't properly retain their owner when reloading the world